requirements.txt
Flask==2.0.0
psycopg2-binary==2.9.1
kubernetes==26.1.0